pyproject.toml
[project]
name = "gitlab-group-exporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.12.0",
    "python-gitlab>=7.0.0",
    "rich>=14.2.0",
    "typer>=0.20.0",
]

[project.scripts]
gitlab-export-group = "gitlab_group_exporter:cli.cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "N", # pep-8 naming
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "C4", # flake-comprehensions
    "SIM", # flake8-simplify
    "I", # isort
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
    "ty>=0.0.1a29",
]
